+++ test/runtests.jl
@@ -25,6 +25,7 @@ end
     include("bitindices.jl")
     include("operations.jl")
     include("determinant.jl")
+    include("factorization.jl")
     include("theorems.jl")
     include("custom_signature.jl")
     include("ext/Adapt.jl")